requirements.txt
gevent
mesos.interface